fix(swagger): resolve ParameterRow crash and enable proper API documentation
- Add complete schema definition to /declarations route with querystring validation and response structure
- Fix route path from '/api/declarations' to '/declarations' to prevent double prefix (/api/api/...)
- Correct JSON Schema format: required field must be array [], not boolean or omitted
- Remove invalid exposeRoute option from @fastify/swagger config (belongs to swagger-ui plugin)
- Add additionalProperties: true to response items schema to support dynamic JOIN fields
- Fix nullable type syntax for OpenAPI 3.0: use { type: 'number', nullable: true } instead of array type
- Add field descriptions to querystring schema for better Swagger UI tooltips
- Add example response to help users understand expected output format

Result: Swagger UI now renders Parameters section without crashes, shows all filter fields with descriptions, and displays proper example responses.

+++ src/modules/declarations/schema.ts
@@ -9,7 +9,7 @@ export const declarationQuerySchema = z.object({
   // Курсор принимает ID последней записи, которую клиент получил на предыдущем шаге.
   // При первом запросе параметр отсутствует, сервер вернёт самые свежие записи.
   cursor: z.coerce.number().int().positive().optional(),
-  
+
   // Сортировка: разрешаем только явно указанные поля (защита от инъекций в ORDER BY)
   //sort_by: z.enum(['id', 'decl_number', 'decl_reg_date', 'update_at']).default('id'),
   sort_order: z.enum(['ASC', 'DESC']).default('DESC'),
@@ -19,8 +19,8 @@ export const declarationQuerySchema = z.object({
   manufacturer_inn: z.string().optional(),
 
   // Фильтры по числам (связанные справочники)
-  status_id: z.coerce.number().int().optional(),
-  doc_type_id: z.coerce.number().int().optional(),
+  status_id: z.coerce.number().optional(),
+  doc_type_id: z.coerce.number().optional(),
 
   // Фильтры по диапазону дат
   reg_date_from: z.iso.date().optional(),
